Initial functionality: load routes from file tree
Initial functionality built.  We can now load routes from modules
living in .js files in a directory (and subdirectories thereof)
specified in the configuration given to us, or './endpoints' by
default if non is specified in the configuration.  We load each
module's specifically defined function and add a route to the
restify instance using the method, route, and handler provided.