Module has multiple functions exposed now
Converting module from returning a single function that sets up a
restify server to instead return that function as one property of
the module, and adding additional functions available as well.
Specifically, we now also have a streamResponse method that
properly wraps a stream in a json-stream-response stream that
behaves properly with restify.