jsonResponder standard node callback factory
Added the jsonResponder standard node callback factory function.
Given the response object, and the next argument to a restify
handler, returns a standard node callback handler (2 arguments, err
and in this case the result to be returned) that checks err and
properly passes it along to next in restify style, otherwise it
sends the result with status 200 in the response.