Implement loadModules mandatory authentication spec
Now loadModules requires all definitions to specify which
authorization/authentication module(s) should be used to control
access.  This includes routes which should have open access, for
which a default "anonymous" authentication method is provided.

Logic is that default behavior is anyone defining endpoints must
make sure that they consider what access control they want on it,
but to make it easy for them to decide that open access should be
allowed.

In addition, the ability to use Passport and its authentication
strategies is accessible now as well.  The artifacts that now
control authentication/authorization are:
- In the definition of routes loaded up via configureServer, each
route must now have an authMethod property defined on it specifying
a correctly loaded authentication/authorization middleware
- The configuration object passed to configureServer must have an
authMethods property on it which contains string keys which are
the names of valid restify middleware for handling
authentication/authorization that are available, and the value is
the middleware function itself
- A "sane" default initializer for the authMethods object that
defines just the "anonymous" authentication method (no checking is
done) on an object and returns it is available as
initAuthWithAnonymous
- There is a loadPassportStrategy method exported as well that
can be passed
    - an object ready to be used as the authMethods property passed
    as the config to configureServer,
    - A passport strategy object, typically created by calling
    the passport strategy module's constructor method with whatever
    arguments it requires (e.g. an options object, a callback method
    for returning whether authorization has been granted)
    - An optional "friendly" name that will be used as the property
    name on the authMethods object instead of whatever the Passport
    strategy itself wants to be called.
	-